+++ Interfaces/ISortAlgorithm.cs
@@ -1,0 +1,10 @@
+using SortingVisualizer.Models;
+using System.Threading;
+
+namespace SortingVisualizer.Interfaces
+{
+    internal interface ISortAlgorithm
+    {
+        void Sort(SortElement[] elements, CancellationToken token);
+    }
+}