+++ Utility/InsertionSort.cs
@@ -7,32 +7,32 @@ using System.Threading;
 
 namespace SortingVisualizer.Utility
 {
-    internal class InsertionSort : IInsertionSort
+    internal class InsertionSort : ISortAlgorithm
     {
-        public void InsertionSortProcess(SortElement[] heights, CancellationToken token)
+        public void Sort(SortElement[] elements, CancellationToken token)
         {
-            for(int i = 1; i < heights.Length; i++)
+            for(int i = 1; i < elements.Length; i++)
             {
                 if (token.IsCancellationRequested) return;
-                int temp = heights[i].Value;
+                int temp = elements[i].Value;
                 int j = i - 1;
-                while(j >= 0 && heights[j].Value > temp)
+                while(j >= 0 && elements[j].Value > temp)
                 {
                     if (token.IsCancellationRequested) return;
-                    Global.Canvas.drawRect(Color.Red, heights[j + 1].Id * Global.Width, Global.MaxHeight - heights[j + 1].Value);
-                    Global.Canvas.drawRect(Color.Red, heights[j].Id * Global.Width, Global.MaxHeight - heights[j].Value);
+                    Global.Canvas.drawRect(Color.Red, elements[j + 1].Id * Global.Width, Global.MaxHeight - elements[j + 1].Value);
+                    Global.Canvas.drawRect(Color.Red, elements[j].Id * Global.Width, Global.MaxHeight - elements[j].Value);
                     Thread.Sleep(15);
-                    Global.Canvas.drawRect(Color.White, heights[j + 1].Id * Global.Width, 0);
-                    Global.Canvas.drawRect(Color.White, heights[j].Id * Global.Width, 0);
-                    heights[j + 1].Value = heights[j].Value;
-                    Global.Canvas.drawRect(Color.Blue, heights[j + 1].Id * Global.Width, Global.MaxHeight - heights[j + 1].Value);
-                    Global.Canvas.drawRect(Color.Blue, heights[j].Id * Global.Width, Global.MaxHeight - heights[j].Value);
+                    Global.Canvas.drawRect(Color.White, elements[j + 1].Id * Global.Width, 0);
+                    Global.Canvas.drawRect(Color.White, elements[j].Id * Global.Width, 0);
+                    elements[j + 1].Value = elements[j].Value;
+                    Global.Canvas.drawRect(Color.Blue, elements[j + 1].Id * Global.Width, Global.MaxHeight - elements[j + 1].Value);
+                    Global.Canvas.drawRect(Color.Blue, elements[j].Id * Global.Width, Global.MaxHeight - elements[j].Value);
                     Thread.Sleep(15);
-                    Global.Canvas.drawRect(Color.Black, heights[j + 1].Id * Global.Width, Global.MaxHeight - heights[j + 1].Value);
-                    Global.Canvas.drawRect(Color.Black, heights[j].Id * Global.Width, Global.MaxHeight - heights[j].Value);
+                    Global.Canvas.drawRect(Color.Black, elements[j + 1].Id * Global.Width, Global.MaxHeight - elements[j + 1].Value);
+                    Global.Canvas.drawRect(Color.Black, elements[j].Id * Global.Width, Global.MaxHeight - elements[j].Value);
                     j--;
                 }
-                heights[j + 1].Value = temp;
+                elements[j + 1].Value = temp;
             }
         }
     }